Complete second rename
